#' Get data

Get data from an elic_cat object.

Usage

cat_get_data(x, mechanism, ..., site = "all")

Arguments

x

an object of class elic_cat.

mechanism

character string that indicates the machanism to which the data belongs.

...

Unused arguments, included only for future extensions of the function.

site

character string with the name of the site or character vector with the sites that you want to extract from the data. Use all for all sites.

Value

A tibble with the extracted data.

See also

Author

Sergio Vignali

Examples

# Create the elic_cat object for an elicitation process with three
# mechanisms, four sites, five levels and a maximum of six experts per
# mechanism
my_levels <- c("level_1", "level_2", "level_3", "level_4", "level_5")
my_sites <- c("site_1", "site_2", "site_3", "site_4")
my_mechanisms <- c("mechanism_1", "mechanism_2", "mechanism_3")
my_elicit <- cat_start(levels = my_levels,
                       sites = my_sites,
                       experts = 6,
                       mechanisms = my_mechanisms) |>
  cat_add_data(data_source = mechanism_1, mechanism = "mechanism_1") |>
  cat_add_data(data_source = mechanism_2, mechanism = "mechanism_2") |>
  cat_add_data(data_source = mechanism_3, mechanism = "mechanism_3")
#> ✔ <elic_cat> object for "Elicitation" correctly initialised
#> ✔ Data added to Mechanism "mechanism_1" from "data.frame"
#> ✔ Data added to Mechanism "mechanism_2" from "data.frame"
#> ✔ Data added to Mechanism "mechanism_3" from "data.frame"

# Get all data from Mechanism 1
cat_get_data(my_elicit, mechanism = "mechanism_1")
#> # A tibble: 120 × 5
#>    id      level   site   confidence estimate
#>    <chr>   <chr>   <chr>       <dbl>    <dbl>
#>  1 5ac97e0 level_1 site_1         15     0.08
#>  2 5ac97e0 level_2 site_1         15     0   
#>  3 5ac97e0 level_3 site_1         15     0.85
#>  4 5ac97e0 level_4 site_1         15     0.02
#>  5 5ac97e0 level_5 site_1         15     0.05
#>  6 5ac97e0 level_1 site_2         35     0.02
#>  7 5ac97e0 level_2 site_2         35     0.11
#>  8 5ac97e0 level_3 site_2         35     0.18
#>  9 5ac97e0 level_4 site_2         35     0.02
#> 10 5ac97e0 level_5 site_2         35     0.67
#> # ℹ 110 more rows

# Get data by site name----
# Get data for site_1 from Mechanism 2
cat_get_data(my_elicit, mechanism = "mechanism_2", site = "site_1")
#> # A tibble: 25 × 5
#>    id      level   site   confidence estimate
#>    <chr>   <chr>   <chr>       <dbl>    <dbl>
#>  1 e51202e level_1 site_1        100     0.09
#>  2 e51202e level_2 site_1        100     0.21
#>  3 e51202e level_3 site_1        100     0.11
#>  4 e51202e level_4 site_1        100     0.59
#>  5 e51202e level_5 site_1        100     0   
#>  6 e78cbf4 level_1 site_1         75     0.31
#>  7 e78cbf4 level_2 site_1         75     0.27
#>  8 e78cbf4 level_3 site_1         75     0.09
#>  9 e78cbf4 level_4 site_1         75     0.17
#> 10 e78cbf4 level_5 site_1         75     0.16
#> # ℹ 15 more rows

# Get data for site_1 and site_3 from Mechanism 3
cat_get_data(my_elicit,
             mechanism = "mechanism_3",
             site = c("site_1", "site_3"))
#> # A tibble: 60 × 5
#>    id      level   site   confidence estimate
#>    <chr>   <chr>   <chr>       <dbl>    <dbl>
#>  1 5ac97e0 level_1 site_1         80     0.02
#>  2 5ac97e0 level_2 site_1         80     0.02
#>  3 5ac97e0 level_3 site_1         80     0.01
#>  4 5ac97e0 level_4 site_1         80     0.87
#>  5 5ac97e0 level_5 site_1         80     0.08
#>  6 5ac97e0 level_1 site_3         15     0.59
#>  7 5ac97e0 level_2 site_3         15     0.08
#>  8 5ac97e0 level_3 site_3         15     0.11
#>  9 5ac97e0 level_4 site_3         15     0   
#> 10 5ac97e0 level_5 site_3         15     0.22
#> # ℹ 50 more rows