Start elicitation

elic_cat_start() initialises an elic_cat object which stores important metadata for the data collected during the elicitation process of categorical data.

Usage

elic_cat_start(
  levels,
  sites,
  experts,
  mechanisms,
  ...,
  title = "Elicitation",
  verbose = TRUE
)

Arguments

levels

character vector with the names of the levels of impact. See Levels for more.

sites

character vector with the names of all sites investigated. See Sites for more.

experts

numeric, indicating the maximum number of experts participating in the elicitation process for one mechanism. See Experts for more.

mechanisms

character vector with the name of the mechanisms of impact investigated.

...

Unused arguments, included only for future extensions of the function.

title

character, used to bind a name to the object.

verbose

logical, if TRUE prints informative messages.

Value

An object of class elic_cat binding metadata related to the elicitation process. These metadata are used by other functions to validate the correctness of the provided data.

Levels

Levels are inherited between mechanisms. A minimum of two levels are needed. If only one level is investigated, please refer to the functions for the elicitation of continuous data (e.g. elic_cont_start).

Sites

The site parameter is a character vector containing the names of all the sites investigated in the elicitation. However, not all sites have to be investigated in every mechanism of impact.

Experts

The expert parameter is a number indicating the maximum number of experts taking part in the elicitation of one of the investigated mechanisms of impact. The number and IDs of experts can differ between the mechanisms.

References

Hemming, V., Burgman, M. A., Hanea, A. M., McBride, M. F., & Wintle, B. C. (2018). A practical guide to structured expert elicitation using the IDEA protocol. Methods in Ecology and Evolution, 9(1), 169–180. https://doi.org/10.1111/2041-210X.12857 Vernet, M., Trask, A.E., Andrews, C.E., Ewen, J.E., Medina, S., Moehrenschlager, A. & Canessa, S. (2024) Assessing invasion risks using EICAT‐based expert elicitation: application to a conservation translocation. Biological Invasions, 26(8), 2707–2721. https://doi.org/10.1007/s10530-024-03341-2

Author

Sergio Vignali and Maude Vernet

Examples

# Create the elic_cat object for an elicitation process over 2 mechanisms of
# impact, 3 sites, 3 levels per sites, and a maximum number of experts per
# mechanism of 8.
my_elicit <- elic_cat_start(levels = c("Level_1", "Level_2", "Level_3"),
                            sites = c("Site_1", "Site_2", "Site_3"),
                            experts = 8,
                            mechanisms = c("mechanism_1","mechanism_2"))
#> ✔ <elic_cat> object for "Elicitation" correctly initialised
my_elicit
#> 
#> ── Elicitation ──
#> 
#> • Levels: "Level_1", "Level_2", and "Level_3"
#> • Sites: "Site_1", "Site_2", and "Site_3"
#> • Number of experts: 8
#> • Mechanismss: "mechanism_1" and "mechanism_2"
#> • Data available for 0 mechanisms

# A title can be added to bind a name to the object:
my_elicit <- elic_cat_start(levels = c("Level_1", "Level_2", "Level_3"),
                            sites = c("Site_1", "Site_2", "Site_3"),
                            experts = 8,
                            mechanisms = c("mechanism_1","mechanism_2"),
                            title = "My elicitation")
#> ✔ <elic_cat> object for "My elicitation" correctly initialised
my_elicit
#> 
#> ── My elicitation ──
#> 
#> • Levels: "Level_1", "Level_2", and "Level_3"
#> • Sites: "Site_1", "Site_2", and "Site_3"
#> • Number of experts: 8
#> • Mechanismss: "mechanism_1" and "mechanism_2"
#> • Data available for 0 mechanisms